+++ src/string.c
@@ -2,8 +2,8 @@
 
 #include "string.h"
 
-uint32_t strlen(const uint8_t* str) {
-    uint32_t len;
+size_t strlen(const uint8_t* str) {
+    size_t len;
     for(len = 0; str[len] != 0; len++) {;}
     return len;
 }
